nest legacy Type components under Type vs TypeBase
