+++ src/components/Type/TitleSmall.js
@@ -1,36 +1,36 @@
-import { TypeBase, TypeFoundry } from './TypeBase.js'
+import { TypeBase, TypeFoundry } from '../TypeBase/TypeBase.js'
 
-export const TitleLarge = {
+export const TitleSmall = {
   Sans: {
     Light300: TypeFoundry({
-      subtype: TypeBase.SUBTYPES.TITLE_LARGE,
+      subtype: TypeBase.SUBTYPES.TITLE_SMALL,
       typeface: TypeBase.TYPEFACES.THEINHARDT,
       weight: TypeBase.WEIGHTS.LIGHT_300,
     }),
-    Medium500: TypeFoundry({
-      subtype: TypeBase.SUBTYPES.TITLE_LARGE,
-      typeface: TypeBase.TYPEFACES.THEINHARDT,
-      weight: TypeBase.WEIGHTS.MEDIUM_500,
-    }),
     Regular400: TypeFoundry({
-      subtype: TypeBase.SUBTYPES.TITLE_LARGE,
+      subtype: TypeBase.SUBTYPES.TITLE_SMALL,
       typeface: TypeBase.TYPEFACES.THEINHARDT,
       weight: TypeBase.WEIGHTS.REGULAR_400,
+    }),
+    Medium500: TypeFoundry({
+      subtype: TypeBase.SUBTYPES.TITLE_SMALL,
+      typeface: TypeBase.TYPEFACES.THEINHARDT,
+      weight: TypeBase.WEIGHTS.MEDIUM_500,
     }),
   },
   Serif: {
     Regular400: TypeFoundry({
-      subtype: TypeBase.SUBTYPES.TITLE_LARGE,
+      subtype: TypeBase.SUBTYPES.TITLE_SMALL,
       typeface: TypeBase.TYPEFACES.CAMBON,
       weight: TypeBase.WEIGHTS.REGULAR_400,
     }),
     Book500: TypeFoundry({
-      subtype: TypeBase.SUBTYPES.TITLE_LARGE,
+      subtype: TypeBase.SUBTYPES.TITLE_SMALL,
       typeface: TypeBase.TYPEFACES.CAMBON,
       weight: TypeBase.WEIGHTS.BOOK_500,
     }),
     Demi600: TypeFoundry({
-      subtype: TypeBase.SUBTYPES.TITLE_LARGE,
+      subtype: TypeBase.SUBTYPES.TITLE_SMALL,
       typeface: TypeBase.TYPEFACES.CAMBON,
       weight: TypeBase.WEIGHTS.DEMI_600,
     }),
