+++ src/components/Type/TitleXLarge.js
@@ -1,36 +1,36 @@
-import { TypeBase, TypeFoundry } from './TypeBase.js'
+import { TypeBase, TypeFoundry } from '../TypeBase/TypeBase.js'
 
-export const TitleMedium = {
+export const TitleXLarge = {
   Sans: {
     Light300: TypeFoundry({
-      subtype: TypeBase.SUBTYPES.TITLE_MEDIUM,
+      subtype: TypeBase.SUBTYPES.TITLE_XLARGE,
       typeface: TypeBase.TYPEFACES.THEINHARDT,
       weight: TypeBase.WEIGHTS.LIGHT_300,
     }),
     Medium500: TypeFoundry({
-      subtype: TypeBase.SUBTYPES.TITLE_MEDIUM,
+      subtype: TypeBase.SUBTYPES.TITLE_XLARGE,
       typeface: TypeBase.TYPEFACES.THEINHARDT,
       weight: TypeBase.WEIGHTS.MEDIUM_500,
     }),
     Regular400: TypeFoundry({
-      subtype: TypeBase.SUBTYPES.TITLE_MEDIUM,
+      subtype: TypeBase.SUBTYPES.TITLE_XLARGE,
       typeface: TypeBase.TYPEFACES.THEINHARDT,
       weight: TypeBase.WEIGHTS.REGULAR_400,
     }),
   },
   Serif: {
     Regular400: TypeFoundry({
-      subtype: TypeBase.SUBTYPES.TITLE_MEDIUM,
+      subtype: TypeBase.SUBTYPES.TITLE_XLARGE,
       typeface: TypeBase.TYPEFACES.CAMBON,
       weight: TypeBase.WEIGHTS.REGULAR_400,
     }),
     Book500: TypeFoundry({
-      subtype: TypeBase.SUBTYPES.TITLE_MEDIUM,
+      subtype: TypeBase.SUBTYPES.TITLE_XLARGE,
       typeface: TypeBase.TYPEFACES.CAMBON,
       weight: TypeBase.WEIGHTS.BOOK_500,
     }),
     Demi600: TypeFoundry({
-      subtype: TypeBase.SUBTYPES.TITLE_MEDIUM,
+      subtype: TypeBase.SUBTYPES.TITLE_XLARGE,
       typeface: TypeBase.TYPEFACES.CAMBON,
       weight: TypeBase.WEIGHTS.DEMI_600,
     }),
