+++ src/components/Type/index.js
@@ -4,8 +4,8 @@ export { Caption } from './../Caption.js'
 export { Footnote } from './../Footnote.js'
 export { Body } from '../Body.js'
 export { Link } from '../Link.js'
-export { TitleSmall } from '../TypeBase/TitleSmall.js'
-export { TitleMedium } from '../TypeBase/TitleMedium.js'
-export { TitleLarge } from '../TypeBase/TitleLarge.js'
-export { TitleXLarge } from '../TypeBase/TitleXLarge.js'
-export { TitleXXLarge } from '../TypeBase/TitleXXLarge.js'
+export { TitleSmall } from './TitleSmall.js'
+export { TitleMedium } from './TitleMedium.js'
+export { TitleLarge } from './TitleLarge.js'
+export { TitleXLarge } from './TitleXLarge.js'
+export { TitleXXLarge } from './TitleXXLarge.js'
