+++ src/components/index.d.ts
@@ -835,11 +835,11 @@ export { Caption } from './Caption.js'
 export { Footnote } from './Footnote.js'
 export { Body } from './Body.js'
 export { Link } from './Link.js'
-export { TitleSmall } from './TypeBase/TitleSmall.js'
-export { TitleMedium } from './TypeBase/TitleMedium.js'
-export { TitleLarge } from './TypeBase/TitleLarge.js'
-export { TitleXLarge } from './TypeBase/TitleXLarge.js'
-export { TitleXXLarge } from './TypeBase/TitleXXLarge.js'
+export { TitleSmall } from './Type/TitleSmall.js'
+export { TitleMedium } from './Type/TitleMedium.js'
+export { TitleLarge } from './Type/TitleLarge.js'
+export { TitleXLarge } from './Type/TitleXLarge.js'
+export { TitleXXLarge } from './Type/TitleXXLarge.js'
 
 // Type2020
 export { TitleSmall2020 } from './Type2020/TitleSmall2020.js'
